Процесс тестирования

+++ src/models/TestResult.java
@@ -4,13 +4,13 @@ import java.sql.Date;
 
 public class TestResult {
     private final long id;
-    private final int idTest;
+    private final long idTest;
     private Test test;
     private final String person;
     private final Date date;
     private final int result;
 
-    public TestResult(long id, int idTest, String person, Date date, int result) {
+    public TestResult(long id, long idTest, String person, Date date, int result) {
         this.id = id;
         this.idTest = idTest;
         this.person = person;
@@ -22,7 +22,7 @@ public class TestResult {
         return id;
     }
 
-    public int getIdTest() {
+    public long getIdTest() {
         return idTest;
     }
